This is several commits in one as i kept forgetting to commit.
The menu has been redesigned to be prettier. it is also possible to return to the menu form the game and back without creating issues. players can exit the game from the menu as well.
A pause menu has been created which will allow players to pause the game and also return to menu.
sound manager has been added for playing can cleaning up sound objects. sounds have been added to player jump, air jump and collecting bits.
players collision detection for being on the ground has been changed from a small circle in the middle to s box along the bottom. The box extends below the player a bit to account for tile collisions bumping the player up, which would cause their state to change from ground to air.
wallslide now accelerates while the player is on the wall until they are at full fall speed.
checkpoints can now be placed as a field, changing the players spawn to an attached point when entered.
changed ladders to hold while the player is not pressing anything, and pressing s (down) now lets the player go down.